KAFKA-4775; Fix findbugs warnings in kafka-tools
Author: Colin P. Mccabe <cmccabe@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #2559 from cmccabe/KAFKA-4775

+++ tools/src/main/java/org/apache/kafka/tools/VerifiableProducer.java
@@ -235,7 +235,7 @@ public class VerifiableProducer {
     /** Returns a string to publish: ether 'valuePrefix'.'val' or 'val' **/
     public String getValue(long val) {
         if (this.valuePrefix != null) {
-            return String.format("%d.%d", this.valuePrefix.intValue(), val);
+            return String.format("%d.%d", this.valuePrefix, val);
         }
         return String.format("%d", val);
     }
